FaceAuthWebApp/requirements.txt
numpy
deepface
pillow
flask